# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API Documentation'
  description: "Documentation des API de l'application Les Innovateurs. Cette documentation fournit des informations détaillées sur les différentes API disponibles pour gérer les utilisateurs, les élections, les candidats, les activités, les sondages et d'autres fonctionnalités de l'application."
  license:
    name: 'URL Accueil'
    url: 'https://lesinnovateurs.me'
  version: 1.0.0
paths:
...
  '/api/public/commune-posts/{commune}':
    get:
      tags:
        - 'Public API'
      summary: "Récupérer tous les posts des d'une commune"
      operationId: getAllPostByCommune
      parameters:
        -
          name: commune
          in: path
          description: Commune
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Liste des posts récupérée avec succès'
        401:
          description: 'Non autorisé'
...